requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
openai==1.30.0
streamlit==1.35.0
python-dotenv==1.0.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pgvector==0.2.5
alembic==1.13.1
pypdf==4.2.0
pydantic==2.7.0
requests==2.32.0
groq==0.9.0
sentence-transformers==3.0.0